Python API/CRUD API/requirements.txt
fastapi
uvicorn
pydantic
passlib[bcrypt]
pyjwt
python-jose